🔨 [#053] - Rename -use-* hooks, extract attendance page subcomponents, and fix test imports 🧹
- 🔄 Renamed 6 hook files in components/employees/ from -use-* to use-* per CLAUDE.md convention
- 📄 Updated all imports across 13 files to reference renamed hook paths
- 🗂️ Extracted SchedulePanel and schedule state components into co-located files under $employeeId/
- ✅ Added unauthenticated_request_returns_401 test with app('auth')->forgetGuards() reset
- 🔧 Fixed stale -use-* import paths in 4 test files (weekly-calendar, use-schedule-section, use-create-day-override, use-create-schedule-inline)
