+++ code/webapp/src/components/employees/override-scope-dialog.tsx
@@ -3,7 +3,7 @@ import { X, Zap, AlertTriangle } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import type { ScheduleDayOverride } from '@/types/schedule'
-import type { OverrideScope } from './-use-create-day-override'
+import type { OverrideScope } from './use-create-day-override'
 import { useOverrideScopeDialog } from './use-override-scope-dialog'
 import { calcDayHours, formatHours, overrideDateLabel } from './schedule-section-utils'
 import { formatTime } from '@/lib/time-format'
