+++ code/webapp/src/components/employees/schedule-content.tsx
@@ -1,7 +1,7 @@
 import { DAY_LABELS } from '@/types/schedule'
 import type { EmployeeSchedule } from '@/types/schedule'
-import type { EditDayValues } from './-use-create-day-override'
-import { useScheduleContent } from './-use-schedule-content'
+import type { EditDayValues } from './use-create-day-override'
+import { useScheduleContent } from './use-schedule-content'
 import { ReadRow, EditRow } from './schedule-config-rows'
 import { OverrideScopeDialog } from './override-scope-dialog'
 import { WeeklyCalendar, OverrideListDialog } from './weekly-calendar'
