+++ code/webapp/src/components/employees/schedule-dialog.tsx
@@ -2,8 +2,8 @@ import { createPortal } from 'react-dom'
 import { CalendarDays, X, Plus, ArrowLeft } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 import type { Employee } from '@/types/employee'
-import type { useScheduleSection } from './-use-schedule-section'
-import { useDialogAnimation } from './-use-dialog-animation'
+import type { useScheduleSection } from './use-schedule-section'
+import { useDialogAnimation } from './use-dialog-animation'
 import { CreateScheduleForm } from './create-schedule-form'
 import { ScheduleContent } from './schedule-content'
 import { EmptySchedule, ScheduleSkeleton } from './schedule-section-state'
