+++ code/webapp/src/components/employees/use-dialog-animation.ts
