+++ code/webapp/src/components/employees/weekly-calendar.tsx
@@ -2,8 +2,8 @@ import { createPortal } from 'react-dom'
 import { X, Zap, ChevronLeft, ChevronRight } from 'lucide-react'
 import { DAY_LABELS } from '@/types/schedule'
 import type { EmployeeSchedule, ScheduleDayOverride } from '@/types/schedule'
-import { resolveWeek, addDays, fmtDayShort } from './-use-weekly-calendar'
-import type { ResolvedDay } from './-use-weekly-calendar'
+import { resolveWeek, addDays, fmtDayShort } from './use-weekly-calendar'
+import type { ResolvedDay } from './use-weekly-calendar'
 import { calcDayHours, formatHours, overrideDateLabel } from './schedule-section-utils'
 import { formatTime } from '@/lib/time-format'
